requirements.txt
openai>=1.14.0
langchain>=0.1.17
python-dotenv>=1.0.1
streamlit>=1.34.0
